1) Former SInstanceData.teamID is replaced with 4 bytes: // u8  teamIdx // u8  drawFlag // u8  unused // u8  unused 2) Respectively Lua side shaders can enjoy drawFlag field, to selectively draw/skip stuff 3) Fixed nasty const correctness bug, removed harmful assert

+++ rts/Rendering/Units/UnitDrawerData.h
@@ -30,7 +30,7 @@ public:
 	float3 dir;
 
 	int facing; //FIXME replaced with dir-vector just legacy decal drawer uses this
-	int team;
+	uint8_t team;
 	int refCount;
 	int lastDrawFrame;
 };
